con dto

+++ trabajoIntegrador/src/main/java/com/dh/odontologia/service/PacienteService.java
@@ -1,6 +1,7 @@
 package com.dh.odontologia.service;
 
 
+import com.dh.odontologia.exceptions.ResourceNotFoundExceptions;
 import com.dh.odontologia.model.Paciente;
 import com.dh.odontologia.model.dto.PacienteDTO;
 import com.dh.odontologia.repository.IPacienteRepository;
@@ -18,8 +19,9 @@ public class PacienteService implements IPacienteService {
     ObjectMapper mapper;
 
     @Override
-    public void crearPaciente(PacienteDTO pacienteDTO) {
+    public PacienteDTO crearPaciente(PacienteDTO pacienteDTO) {
         guardarPaciente(pacienteDTO);
+        return pacienteDTO;
     }
 
     @Override
@@ -31,28 +33,35 @@ public class PacienteService implements IPacienteService {
         }
         return pacienteDTO;
     }
-    private void guardarPaciente(PacienteDTO pacienteDTO){
+    private  Paciente guardarPaciente(PacienteDTO pacienteDTO){
         Paciente paciente = mapper.convertValue(pacienteDTO,Paciente.class);
-        pacienteRepository.save(paciente);
+        return pacienteRepository.save(paciente);
     }
 
     @Override
-    public void modificarPaciente(PacienteDTO pacienteDTO) {
+    public void modificarPaciente(PacienteDTO pacienteDTO) throws ResourceNotFoundExceptions {
+       if (leerPaciente(pacienteDTO.getId())==null)
+           throw new ResourceNotFoundExceptions("No existe un paciente con el id"+ pacienteDTO.getId());
         guardarPaciente(pacienteDTO);
     }
 
     @Override
-    public void eliminarPaciente(Long id) {
+    public void eliminarPaciente(Long id) throws ResourceNotFoundExceptions {
+        if (leerPaciente(id)==null)
+            throw  new ResourceNotFoundExceptions("No existe el paciente con el id:"+ id);
+
         pacienteRepository.deleteById(id);
     }
 
     @Override
     public Collection<PacienteDTO> listarPacientes() {
+
         List<Paciente> pacientes =pacienteRepository.findAll();
         Set<PacienteDTO> pacientesDTO = new HashSet<PacienteDTO>();
         for (Paciente paciente: pacientes) {
             pacientesDTO.add(mapper.convertValue(paciente,PacienteDTO.class));
         }
+
         return pacientesDTO;
     }
